Minimal realisation should be done still TODO: Add groups and servers Add voice chats Add message reactions Add selecting pinned messages Change error codes

+++ routes/messages_routes.py
@@ -2,7 +2,7 @@ from app import app
 from bson import ObjectId
 from bson import errors as  bson_errors
 from quart import Response, request, jsonify
-from views import UserView
+from views import UserView, users_connected_ids
 from models.endpoint import TextEndpoint
 from .middlewares import (
     validate_api_version,
@@ -17,7 +17,7 @@ from .schemas import (
 )
 
 
-@app.route("/api/users/me/endpoints/<string:endpoint_id>/messages")
+@app.route("/api/endpoints/<string:endpoint_id>/messages")
 @validate_api_version("1.0.0")
 @authorized
 async def get_messages_latest(user: UserView, endpoint_id: str):
@@ -33,14 +33,13 @@ async def get_messages_latest(user: UserView, endpoint_id: str):
 
     try:
         messages = await endpoint.get_messages(user._id, endpoint.last_message)
-        # resend to all members
         return success({"messages":[message.__dict__ for message in messages]})
 
     except TextEndpoint.exc.NotGroupMember:
         return error("you aren't a group member")
 
 
-@app.route("/api/users/me/endpoints/<string:endpoint_id>/messages/<string:message_id>")
+@app.route("/api/endpoints/<string:endpoint_id>/messages/from/<string:message_id>")
 @validate_api_version("1.0.0")
 @authorized
 async def get_messages_from(user: UserView, endpoint_id: str, message_id: str):
@@ -57,7 +56,6 @@ async def get_messages_from(user: UserView, endpoint_id: str, message_id: str):
     try:
         message_id = ObjectId(message_id)
         messages = await endpoint.get_messages(user._id, endpoint.last_message)
-        # resend to all members
         return success({"messages":[message.__dict__ for message in messages]})
 
     except TextEndpoint.exc.NotGroupMember:
@@ -67,7 +65,7 @@ async def get_messages_from(user: UserView, endpoint_id: str, message_id: str):
         return error("Invalid message id")
 
 
-@app.route("/api/users/me/endpoints/<string:endpoint_id>/messages/<string:message_id>")
+@app.route("/api/endpoints/<string:endpoint_id>/messages/<string:message_id>")
 @validate_api_version("1.0.0")
 @authorized
 async def get_message(user: UserView, endpoint_id: str, message_id: str):
@@ -94,7 +92,7 @@ async def get_message(user: UserView, endpoint_id: str, message_id: str):
         return error("Invalid message id")
 
 
-@app.route("/api/users/me/endpoints/<string:endpoint_id>/messages/<string:message_id>", methods=["PATCH"])
+@app.route("/api/endpoints/<string:endpoint_id>/messages/<string:message_id>", methods=["PATCH"])
 @validate_api_version("1.0.0")
 @validate_schema(message)
 @authorized
@@ -113,7 +111,8 @@ async def edit_message(user: UserView, endpoint_id: str, message_id: str):
     try:
         message_id = ObjectId(message_id)
         message_modified = await endpoint.edit_message(user._id, message_id, content)
-        # resend to all members that message updated
+
+        #TODO resend to all members that message updated
         return success({"message_edited": message_modified})
 
     except TextEndpoint.exc.NotGroupMember:
@@ -122,12 +121,13 @@ async def edit_message(user: UserView, endpoint_id: str, message_id: str):
     except bson_errors.InvalidId:
         return error("Invalid message id")
 
-
-@app.route("/api/users/me/endpoints/<string:endpoint_id>/messages/<string:message_id>&pin=<int:pinned>", methods=["PATCH"])
+# TODO: add get pins
+@app.route("/api/endpoints/<string:endpoint_id>/messages/<string:message_id>/pin", methods=["PATCH"])
 @validate_api_version("1.0.0")
 @authorized
-async def set_pin_state_message(user: UserView, endpoint_id: str, message_id: str, pinned: int):
+async def set_pin_state_message(user: UserView, endpoint_id: str, message_id: str):
     try:
+        pinned = request.args.get('accept', 'False') == 'True'
         endpoint_id = ObjectId(endpoint_id)
         endpoint = await user.small_endpoint(endpoint_id)
 
@@ -139,8 +139,9 @@ async def set_pin_state_message(user: UserView, endpoint_id: str, message_id: st
 
     try:
         message_id = ObjectId(message_id)
-        message_pinned = await endpoint.pin_message(user._id, message_id, bool(pinned))
-        # resend to all members that message updated
+        message_pinned = await endpoint.pin_message(user._id, message_id, pinned)
+
+        #TODO resend to all members that message updated
         if message_pinned:
             return success({"message_pinned": message_id})
 
@@ -154,7 +155,7 @@ async def set_pin_state_message(user: UserView, endpoint_id: str, message_id: st
         return error("Invalid message id")
 
 
-@app.route("/api/users/me/endpoints/<string:endpoint_id>/messages", methods=["POST"])
+@app.route("/api/endpoints/<string:endpoint_id>/messages", methods=["POST"])
 @validate_api_version("1.0.0")
 @validate_schema(message)
 @authorized
@@ -172,7 +173,15 @@ async def post_message(user: UserView, endpoint_id: str):
 
     try:
         message = await endpoint.send_message(user._id, **data)
-        # resend to all members
+        for user_id in endpoint.members:
+
+            user: UserView = users_connected_ids.get(user_id, None)
+
+            if not user:
+                continue
+
+            user.message_queue.append(message)
+
         return success({"message":message.__dict__})
 
     except TextEndpoint.exc.NotGroupMember:
@@ -182,11 +191,13 @@ async def post_message(user: UserView, endpoint_id: str):
         return error("Too long message")
 
 
-@app.route("/api/users/me/endpoints/<string:endpoint_id>/messages/<string:message_id>", methods=["DELETE"])
+@app.route("/api/endpoints/<string:endpoint_id>/messages/<string:message_id>", methods=["DELETE"])
 @validate_api_version("1.0.0")
 @authorized
 async def delete_message(user: UserView, endpoint_id: str, message_id: str):
     try:
+        force_delete = bool(request.args.get('force', False))
+
         endpoint_id = ObjectId(endpoint_id)
         endpoint = await user.small_endpoint(endpoint_id)
 
@@ -194,12 +205,15 @@ async def delete_message(user: UserView, endpoint_id: str, message_id: str):
         return error("Invalid endpoint id")
 
     except ValueError:
-        return error("No such endpoint")
+        return error("No such endpoint or incorrect ")
 
     try:
         message_id = ObjectId(message_id)
-        message = await endpoint.delete_message(user._id, message_id)
-        # resend to all members
+        if not force_delete:
+            message = await endpoint.delete_message(user._id, message_id)
+        else:
+            message = await endpoint.force_delete(user._id, message_id)
+        #TODO: resend to all members
         return success({"deleted": message})
 
     except TextEndpoint.exc.NotGroupMember:
@@ -208,32 +222,5 @@ async def delete_message(user: UserView, endpoint_id: str, message_id: str):
     except bson_errors.InvalidId:
         return error("Invalid message id")
 
-
-@app.route("/api/users/me/endpoints/<string:endpoint_id>/messages/<string:message_id>&force", methods=["DELETE"])
-@validate_api_version("1.0.0")
-@authorized
-async def force_delete_message(user: UserView, endpoint_id: str, message_id: str):
-    try:
-        endpoint_id = ObjectId(endpoint_id)
-        endpoint = await user.small_endpoint(endpoint_id)
-
-    except bson_errors.InvalidId:
-        return error("Invalid endpoint id")
-
-    except ValueError:
-        return error("No such endpoint")
-
-    try:
-        message_id = ObjectId(message_id)
-        message = await endpoint.force_delete(user._id, message_id)
-        # resend to all members
-        return success({"deleted": message})
-
-    except TextEndpoint.exc.NotGroupMember:
-        return error("you aren't a group member")
-
-    except bson_errors.InvalidId:
-        return error("Invalid message id")
-
-    except TextEndpoint.exc.UnsupprotedMethod:
-        return error("This method unsupproted by this endpoint")
+    except endpoint.exc.EndpointException:
+        return error("You can't force delete this message")
